pyproject.toml
[tool.poetry]
name = "avaprodetailing-bot-1"
version = "0.1.0"
description = ""
authors = ["Валерий <94367062+gugutsidze-vv@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.4.1"
fastapi = "^0.110.0"
python-dotenv = "^1.0.1"
uvicorn = "^0.29.0"
watchdog = "^4.0.0"
sqlalchemy = "^2.0.29"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
aiosqlite = "^0.20.0"
pydantic-settings = "^2.2.1"
fastapi-users-db-sqlalchemy = "4.0.3"
python-multipart = "^0.0.9"
jinja2 = "^3.1.3"
yookassa = "^3.1.0"
qrcode = {extras = ["pillow"], version = "^7.4.2"}
reportlab = "^4.1.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
mypy = "^1.9.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
